moved classes to right packages
