+++ src/test/java/fi/metatavu/metaform/test/functional/tests/AttachmentTestsIT.java
@@ -1,9 +1,13 @@
-package fi.metatavu.metaform.test.functional;
+package fi.metatavu.metaform.test.functional.tests;
 
 import fi.metatavu.metaform.api.client.models.Reply;
 import fi.metatavu.metaform.server.rest.ReplyMode;
 import fi.metatavu.metaform.test.TestSettings;
+import fi.metatavu.metaform.test.functional.AbstractIntegrationTest;
+import fi.metatavu.metaform.test.functional.FileUploadResponse;
 import fi.metatavu.metaform.test.functional.builder.TestBuilder;
+import fi.metatavu.metaform.test.functional.builder.resources.KeycloakResource;
+import fi.metatavu.metaform.test.functional.builder.resources.MysqlResource;
 import io.quarkus.test.common.QuarkusTestResource;
 import io.quarkus.test.junit.QuarkusTest;
 import org.apache.commons.codec.digest.DigestUtils;
